+++ src/runtime/__tests__/project-policy-domain-change.test.ts
@@ -52,7 +52,7 @@ function createTestRuntimeState(): RuntimeState {
       shape2d: new Uint32Array(0),
     },
     arena: new Float32Array(0),
-    state: new Float64Array(0),
+    state: new Float32Array(0),
     eventScalars: new Uint8Array(0),
     eventPrevPredicateValue: new Uint8Array(0),
     events: new Map(),
